+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#154] Ignore missing CAS dependency if `manager.env.authRegistrationEnabled` is set to `true` in the Helm values.
+- [#154] Ignore missing Postfix dependency if `manager.env.disablePostfixDependencyCheck` is set to `true` in the Helm values.
 
 ## [v3.2.0] - 2026-03-25
 ### Removed